Rename 'recommender/output' to 'recommender/checkpoint'.
